+++ gui/Settings.h
@@ -154,6 +154,8 @@ enum class GuiSettingsId {
 
     REDUCE_LOWFREQUENCY_NOISE,
 
+    BLOOM_RADIUS,
+
     BLOOM_INTENSITY,
 
     SHOW_KEY,